TUSS4470_shield_002/web/pyproject.toml
[project]
name = "open-echo-web"
version = "0.1.0"
description = "Web interface for OpenEcho"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "numpy>=2.3.2",
    "pydantic-settings>=2.10.1",
    "pyserial<3.5",
    "pyserial-asyncio>=0.6",
    "pyserial-asyncio-fast>=0.16",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
